remove provide_data_types and provide_label_types
