+++ python/mxnet/io.py
@@ -3,7 +3,7 @@
 
 """NDArray interface of mxnet"""
 from __future__ import absolute_import
-from collections import OrderedDict
+from collections import OrderedDict, namedtuple
 
 import re
 import sys
@@ -15,6 +15,7 @@ from .base import _LIB
 from .base import c_array, c_str, mx_uint, py_str
 from .base import DataIterHandle, NDArrayHandle
 from .base import check_call, ctypes2docstring
+from .base import mx_real_t
 from .ndarray import NDArray
 from .ndarray import array
 from .ndarray import concatenate
@@ -85,8 +86,7 @@ class DefaultLayoutMapper(LayoutMapper):
 class DataBatch(object):
     """Default object for holding a mini-batch of data and related information."""
     def __init__(self, data, label, pad=None, index=None,
-                 bucket_key=None, provide_data=None, provide_label=None,
-                 provide_data_type=None, provide_label_type=None):
+                 bucket_key=None, provide_data=None, provide_label=None):
         self.data = data
         self.label = label
         self.pad = pad
@@ -96,8 +96,6 @@ class DataBatch(object):
         self.bucket_key = bucket_key
         self.provide_data = provide_data
         self.provide_label = provide_label
-        self.provide_data_type = provide_data_type
-        self.provide_label_type = provide_label_type
 
 class DataIter(object):
     """DataIter object in mxnet. """
@@ -203,9 +201,7 @@ class ResizeIter(DataIter):
         self.current_batch = None
 
         self.provide_data = data_iter.provide_data
-        self.provide_data_type = data_iter.provide_data_type
         self.provide_label = data_iter.provide_label
-        self.provide_label_type = data_iter.provide_label_type
         self.batch_size = data_iter.batch_size
 
     def reset(self):
@@ -306,32 +302,16 @@ class PrefetchingIter(DataIter):
         if self.rename_data is None:
             return sum([i.provide_data for i in self.iters], [])
         else:
-            return sum([[(r[n], s) for n, s in i.provide_data] \
+            return sum([[DataDesc(r[x.name], x.shape, x.dtype) for x in i.provide_data] \
                        for r, i in zip(self.rename_data, self.iters)], [])
 
-    @property
-    def provide_data_type(self):
-        """The name and shape of data type provided by this iterator"""
-        if self.rename_data is None:
-            return sum([i.provide_data_type for i in self.iters], [])
-        else:
-            return sum([[(r[n], s) for n, s in i.provide_data_type] \
-                       for r, i in zip(self.rename_data, self.iters)], [])
     @property
     def provide_label(self):
         """The name and shape of label provided by this iterator"""
         if self.rename_label is None:
             return sum([i.provide_label for i in self.iters], [])
         else:
-            return sum([[(r[n], s) for n, s in i.provide_label] \
-                       for r, i in zip(self.rename_label, self.iters)], [])
-
-    def provide_label_type(self):
-        """The name and shape of label type provided by this iterator"""
-        if self.rename_label is None:
-            return sum([i.provide_label_type for i in self.iters], [])
-        else:
-            return sum([[(r[n], s) for n, s in i.provide_label_type] \
+            return sum([[DataDesc(r[x.name], x.shape, x.dtype) for x in i.provide_label] \
                        for r, i in zip(self.rename_label, self.iters)], [])
 
     def reset(self):
@@ -472,22 +452,18 @@ class NDArrayIter(DataIter):
     @property
     def provide_data(self):
         """The name and shape of data provided by this iterator"""
-        return [(k, tuple([self.batch_size] + list(v.shape[1:]))) for k, v in self.data]
-
-    @property
-    def provide_data_type(self):
-        """The name and shape of data type provided by this iterator"""
-        return [(k, v.dtype) for k, v in self.data]
+        return [
+            DataDesc(k, tuple([self.batch_size] + list(v.shape[1:])), v.dtype)
+            for k, v in self.label
+        ]
 
     @property
     def provide_label(self):
         """The name and shape of label provided by this iterator"""
-        return [(k, tuple([self.batch_size] + list(v.shape[1:]))) for k, v in self.label]
-
-    @property
-    def provide_label_type(self):
-        """The name and shape of label type provided by this iterator"""
-        return [(k, v.dtype) for k, v in self.label]
+        return [
+            DataDesc(k, tuple([self.batch_size] + list(v.shape[1:])), v.dtype)
+            for k, v in self.label
+        ]
 
     def hard_reset(self):
         """Igore roll over data and set to start"""
@@ -546,7 +522,6 @@ class MXDataIter(DataIter):
         # debug option, used to test the speed with io effect eliminated
         self._debug_skip_load = False
 
-
         # load the first batch to get shape information
         self.first_batch = None
         self.first_batch = self.next()
@@ -554,10 +529,8 @@ class MXDataIter(DataIter):
         label = self.first_batch.label[0]
 
         # properties
-        self.provide_data = [(data_name, data.shape)]
-        self.provide_data_type = [(data_name, data.dtype)]
-        self.provide_label = [(label_name, label.shape)]
-        self.provide_label_type = [(label_name, label.dtype)]
+        self.provide_data = [DataDesc(data_name, data.shape, data.dtype)]
+        self.provide_label = [DataDesc(label_name, label.shape, label.dtype)]
         self.batch_size = data.shape[0]
 
 
@@ -628,6 +601,35 @@ class MXDataIter(DataIter):
         pad = ctypes.c_int(0)
         check_call(_LIB.MXDataIterGetPadNum(self.handle, ctypes.byref(pad)))
         return pad.value
+
+# pylint: disable=W0622
+class DataDesc(namedtuple('DataDesc', ['name', 'shape'])):
+    """Named data desc description contains name, shape, type and other extended attributes.
+    """
+    def __new__(cls, name, shape, dtype=mx_real_t, layout='NCHW'):
+        ret = super(cls, DataDesc).__new__(cls, name, shape)
+        ret.dtype = dtype
+        ret.layout = layout
+        return ret
+
+    def __repr__(self):
+        return "DataDesc[%s,%s,%s,%s]" % (self.name, self.shape, self.dtype,
+                                          self.layout)
+
+    @staticmethod
+    def get_list(shapes, types):
+        """Get DataDesc list from attribute lists.
+
+        Parameters
+        ----------
+        shapes : shape tuple list with (name, shape) tuples
+        types : type tuple list with (name, type) tuples
+        """
+        if types is not None:
+            type_dict = dict(types)
+            return [DataDesc(x[0], x[1], type_dict[x[0]]) for x in shapes]
+        else:
+            return [DataDesc(x[0], x[1]) for x in shapes]
 
 def _make_io_iterator(handle):
     """Create an io iterator by handle."""
